lios/services/telegram-collector/requirements.txt
telethon>=1.36
python-dotenv
supabase
google-generativeai
schedule
